+++ automerge-protocol/src/utility_impls/mod.rs
@@ -6,3 +6,5 @@ mod key;
 mod object_id;
 mod opid;
 mod scalar_value;
+mod scalar_value_kind;
+mod scalar_values;
