+++ automerge-protocol/src/utility_impls/scalar_value_kind.rs
@@ -1,0 +1,9 @@
+use std::fmt;
+
+use crate::ScalarValueKind;
+
+impl fmt::Display for ScalarValueKind {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
